requirements.txt
localflavor
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-autoslug==1.9.6
django-crispy-forms==1.9.0
django-localflavor==3.0.1
django-mobile==0.7.0
idna==2.9
oauthlib==3.1.0
Pillow==7.1.1
python-stdnum==1.13
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
stripe==2.45.0
urllib3==1.25.8
